Added a new module to ease development.
This required adding the more sophisticated memoize library (more than
Clojure's built-in memoize, that is). This allows us to conditionally purge the
memoize cache for a given function (in this case, the function that reads the
project.clj file).

Other changes include: README update for using this and a new util/in?
function.