Implement find-rod behavior.
1. Use GDAL inverse tranformation to turn projection coordinates into tile-relative raster grid coordinates.
2. Extract appropriate substring from string encoded data and replace data in result with subset of data. This preserves key metadata like the acquisition data and source scene ID.

TBD: I have not decided what to do with the tile's (x,y) value. It is currently still the tile's upper-left (x,y) ... but this seems very likely to cause confusion. The point could be updated instead to match the requested point, and the original tile (x,y) could be associated with different keys like tile-x and tile-y...