Fixed issue with running Docker images.
In truth, this code is a partial fix -- the other part of the fix actually
involved updating the sudoers file in /etc, moving customizations (for running
docker without passwords) into /etc/sudoers.d/, since the admin level config in
/etc/sudoers was overriding the customizations and passwords were still
required (and silently being ignored, due to the fact that the code wasn't
accounting for data in stderr).

The code change in this commit will now return errors from executing to the
shell, so things shouldn't be hidden anymore.