Update find-tile
The previous approach returned a full spec for each tile. This ends up resending a lot of the same data since the spec is the same for all tiles!

Commented out the find-rod function. Clients are capable of extracting rods from tiles. If this ends up being a more common use case, we'll need to fully implement find-rod... but this opens up a can of worms related to encoding and how fill/scale/masking are handled.